+++ src/cjs/PlotBoilerplate.js
@@ -74,7 +74,8 @@
  * @modified 2021-03-03 Added the `VEllipseSector` drawable.
  * @modified 2021-03-29 Clearing `currentClassName` and `currentId` after drawing each drawable.
  * @modified 2021-04-25 Extending `remove` to accept arrays of drawables.
- * @version  1.14.0
+ * @modified 2021-11-16 Adding the `PBText` drawable.
+ * @version  1.15.0
  *
  * @file PlotBoilerplate
  * @fileoverview The main class.
@@ -102,6 +103,7 @@ var VEllipseSector_1 = require("./VEllipseSector");
 var Vector_1 = require("./Vector");
 var Vertex_1 = require("./Vertex");
 var VertexAttr_1 = require("./VertexAttr");
+var PBText_1 = require("./PBText");
 /**
  * @classdesc The main class of the PlotBoilerplate.
  *
@@ -340,6 +342,12 @@ var PlotBoilerplate = /** @class */ (function () {
             image: {
                 color: "#a8a8a8",
                 lineWidth: 1
+            },
+            text: {
+                color: "rgba(192,0,128,0.5)",
+                lineWidth: 1,
+                fill: true,
+                anchor: true
             }
         }; // END drawConfig
         // +---------------------------------------------------------------------------------
@@ -647,6 +655,11 @@ var PlotBoilerplate = /** @class */ (function () {
             });
             drawable.lowerRight.attr.selectable = false;
         }
+        else if (drawable instanceof PBText_1.PBText) {
+            this.vertices.push(drawable.anchor);
+            this.drawables.push(drawable);
+            drawable.anchor.attr.selectable = false;
+        }
         else {
             throw "Cannot add drawable of unrecognized type: " + typeof drawable + ".";
         }
@@ -744,6 +757,9 @@ var PlotBoilerplate = /** @class */ (function () {
                     else if (drawable instanceof PBImage_1.PBImage) {
                         this.removeVertex(drawable.upperLeft, false);
                         this.removeVertex(drawable.lowerRight, false);
+                    }
+                    else if (drawable instanceof PBText_1.PBText) {
+                        this.removeVertex(drawable.anchor, false);
                     }
                 } // END removeWithVertices
                 if (redraw) {
@@ -1069,6 +1085,16 @@ var PlotBoilerplate = /** @class */ (function () {
                 draw.circleHandle(d.lowerRight, 3, this.drawConfig.image.color);
                 d.lowerRight.attr.renderTime = renderTime;
             }
+        }
+        else if (d instanceof PBText_1.PBText) {
+            fill.setCurrentId(d.uid);
+            fill.text(d.text, d.anchor.x, d.anchor.y, d);
+            if (this.drawConfig.text.anchor) {
+                draw.setCurrentId(d.uid + "_a0");
+                draw.setCurrentClassName(d.className + "-anchor");
+                (this.drawConfig.text.fill ? fill : draw).point(d.anchor, this.drawConfig.text.color);
+            }
+            d.anchor.attr.renderTime = renderTime;
         }
         else {
             console.error("Cannot draw object. Unknown class.");