+++ src/ts/interfaces/core.ts
@@ -13,6 +13,7 @@
  * @modified 2021-05-31 Added the `drawLib.setConfiguration` function.
  * @modified 2021-05-31 Splitted the large interfaces.ts file into this one and others.
  * @modified 2021-06-21 Added `IBounds.getCenter()`.
+ * @modified 2021-11-16 Added `text` options to the `DrawConfig`.
  **/
 
 import { Vertex } from "../Vertex";
@@ -28,6 +29,7 @@ import { BezierPath } from "../BezierPath";
 import { Line } from "../Line";
 import { PlotBoilerplate } from "../PlotBoilerplate";
 import { DrawLib } from "./DrawLib";
+import { PBText } from "../PBText";
 
 /**
  * @classdesc Coordinates (x,y) on the plane.
@@ -73,6 +75,7 @@ export type Drawable =
   | Circle
   | CircleSector
   | PBImage
+  | PBText
   | VEllipse
   | VEllipseSector
   | Polygon
@@ -189,6 +192,12 @@ export interface DrawConfig {
   line: DrawSettings;
   vector: DrawSettings;
   image: DrawSettings;
+  text: {
+    color: string;
+    lineWidth: number;
+    fill?: boolean;
+    anchor?: boolean;
+  };
 }
 
 /** This is a fix for the problem, that the constructor's "name" attribute is not